requirements.txt
daphne==4.1.2
django==5.2.3
django-ninja==1.3.0
pre-commit==4.2.0
pydantic==2.11.7
pydantic-settings==2.8.1
psycopg==3.2.6
django-cors-headers
web3==6.10.0
ipfshttpclient==0.8.0a2
celery>=5.3.6
redis>=5.0.4
python-dotenv>=1.0.1
py-solc-x>=1.1.1
ruff>=0.4.4
gunicorn>=22.0.0
psycopg2-binary>=2.9.9
PyPDF2>=3.0.0
setuptools==80.9.0
mnemonic==0.20
django-ratelimit
requests
djangorestframework-simplejwt==5.5.0
py-cid==0.3.0
djangorestframework==3.16.0
